Refactoring: combine action and bool guard in a single module
