+++ ModelExplorerPlugin/PropertyViewBuilderBase.cpp
@@ -3,29 +3,9 @@
 #include "PropertyViewBuilderBase.h"
 #include "PropertyManager.h"
 #include "GuidUtils.h"
+#include "ScopeGuard.h"
 
 #include <Renga/QuantityIds.h>
-
-
-namespace
-{
-  template <typename TFunc>
-  class CActionGuard
-  {
-  public:
-    CActionGuard(const TFunc& func) : m_func(func) {}
-    ~CActionGuard() { m_func(); }
-
-  private:
-    TFunc m_func;
-  };
-
-  template <typename TFunc>
-  CActionGuard<TFunc> makeGuard(const TFunc& func)
-  {
-    return CActionGuard<TFunc>(func);
-  }
-}
 
 
 PropertyViewBuilderBase::PropertyViewBuilderBase(bool disableProperties) : m_disableProperties(disableProperties)
